Resolviendo problema con el Toast.

+++ Diego-P1-AP2/Migrations/20210602235125_Inicial.cs
@@ -13,7 +13,7 @@ namespace Diego_P1_AP2.Migrations
                     ProductoId = table.Column<int>(type: "INTEGER", nullable: false)
                         .Annotation("Sqlite:Autoincrement", true),
                     Descripcion = table.Column<string>(type: "TEXT", nullable: false),
-                    Existencia = table.Column<int>(type: "INTEGER", nullable: false),
+                    Existencia = table.Column<float>(type: "REAL", nullable: false),
                     Costo = table.Column<float>(type: "REAL", nullable: false),
                     ValorInventario = table.Column<float>(type: "REAL", nullable: false)
                 },
